Provide default variables to the template
Any action in the chain (having access to the Context) can now set
variables. For now, the ass2m_version and the webpath are provided to
all actions.
The 'global' dict is just a trick from "So what if I want to set values
that are global to everyone within a template request?" found at
http://www.makotemplates.org/docs/runtime.html#context-variables
The way to handle the responses are simplified, you only have to set the
response body with ctx.res.body = ctx.render(templatename).

Closes #548