Better handling of paths and URLs
path and realpath are relative to the file.
url and relurl are relative to the web application; they keep the query
string.

Use paste's URL class to alter the current URLs. That way, we can for
instance easily rewrite /path/?action=list?view=html
to /path?action=list?view=gallery.