Retrieve the File object once, in the Context
This will allow to make only one call per request to the Storage for the
requested file, and factorize os.path calls (this is still in progress).

There also was a bug: when on the root dir, the path was "/", not "",
however "" is used by the Storage for the root dir, not "/".
This required a few hacks (we want to display "/" instead of "" on the web
even if it is incorrect, and posixpath.join(path, file) had to be changed
to posixpath.join("/", path, file) to be sure to have a starting "/" when
the path is "".
However this last hack should be removed by other improvements to
follow.