Rewriting of the routing system
Views are related to files, actions are not.
The code is much simpler as a result.
This is just the minimum, new features will come up thanks to this.

It also fixes a bug: HEAD requests always returned 404 errors. However
there is no method validation, this might not be ideal.

refs #609