Fix issues with URL escaping and cookies
Even if for the server a quoted or unquoted URL is considered to be the
same, the client will differentiate a cookie if its path is quoted or
not. To work around this horrible bug present in many browsers, I added two things:
* if we can figure out the client-side URL, check if it is fully escaped
 (this will also be useful to force users to have good URLs to
 copy/paste, URLs with spaces in them are always annoying)
* always use the escaped URLs in form actions, redirections, etc. (fixes #633)

Also fix a bug where the cookie could not be unset when the root path was not
'/' (we have to provide a path to delete the cookie too).