Move DATA_PATHS to Storage
This will allow the user to easily overload templates and assets,
for only the current ass2m root.

Another benefit is that we are now able to get the DATA_PATHS without a
Context.

It is possible to get the DATA_PATHS without a directory though, by
getting Storage.DATA_PATHS (this is only useful for errors about the
ass2m root).