Add a way to use HTTP authentication
It is only used when ?authby=http is in the parameters.
The reason is anonymous access and the ability to logout with most
browsers (they keep sending the username/password).
The goal is to be able to use WGet and other simple downloaders, without
requiring cookies.
It is possible to login by key if the username provided is _key.