Add support for hooks
The hooks are designed to be very generic: a name, and corresponding callable.
There can be more than one callable for a single name.
The goal is to help plugins extend key parts of ass2m, like the upcoming
cleanup (fsck/gc) command, or the prewrite/postread steps of
storage objets.