Add empty alt attribute
The alt is left empty, because we already give a description of the
picture in the figcaption tag. Adding an alt content would make screen
readers say it twice.
Since the image and its caption are in the same link tag, we do not have
a textless link either.

fixes #744